Error fix for buying items and coinflip breakdown
Changed coinflip to only need one embed that is edited for each result.

Also added for if they don't respond before the time limit they lose the coin.

Fixed buying an item not working but trying to implement an amount of items to multiply instead of one at a time.